source/requirements.txt
autopep8==1.3.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
dill==0.2.7.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-toolbelt==0.0.1
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.3.0
MarkupSafe==1.0
multiprocessing-on-dill==3.5.0a4
pep8==1.7.0
psycopg2==2.7.3
pyasn1==0.2.3
pycodestyle==2.3.1
pyreadline==2.1
pytz==2017.2
requests==2.18.2
rsa==3.4.2
six==1.10.0
static3==0.7.0
thrift==0.10.0
urllib3==1.22
Werkzeug==0.12.2
Scipy==0.19.1
